+++ src/main/java/com/brigada/bloss/entity/Review.java
@@ -1,4 +1,4 @@
-package com.brigada.bloss.model;
+package com.brigada.bloss.entity;
 
 import java.util.Date;
