--- src/main/java/com/brigada/bloss/model/User.java
@@ -1,20 +1,0 @@
-package com.brigada.bloss.model;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import lombok.ToString;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@ToString
-public class User {
-    
-    private String id;
-    private String name;
-    private String surname;
-    private String login;
-    private String password;
-    
-}
